Adds an endpoint to expose a public JWK.
When implementing a JWK authentication flow (e.g. for LTI Extension),
you need to expose a public JWK that the external service will use to encode
payloads. You need to keep the private key and be able to decode those payloads
(e.g. a JWT id_token) using it. This commit adds the endpoint to expose the public JWK
along with the ability to store and decode payloads using the private key.

Test Plan:
- In the Rails console run Keypair.create! and make sure that when you go to
  https://platformweb/keypairs that you can see the public JWK for it.
- Encode a payload using that pulic JWK and make sure you can decode it with
  Keypair.jwt_decode(encoded_payload)
- Run: bundle exec rspec spec/modles/keypair_spec.rb