First pass at a FormAssembly driven waivers form.
DO NOT MERGE. This is hacked up as a proof of concept. If you insert
the "waivers" when selecting an LTI Assignment in Canvas, when you launch
this it will demonstrate the flow as a hard-coded test participant (and hardcoded test form).

Note: the crazy icons after you hit Sign and Submit depends on this fix:
https://github.com/beyond-z/canvas-lms-js-css/pull/320

Note: none of the actual submit part of the flow is implemented. In the `create`
action we need to use the LtiAdvantageAPI to send a "score" back to Canvas for this
assignment (even through it doesn't count towards your grade).

Test Plan:
- Create an assignment in Canvas, choose External Tool as the submission type.
  Select "Insert Waivers".
- Login as a test student and open the assignment. It should allow you to sign and submit and direct
  you back to Brians Playground course in Canvas at the end.