Only allow editing CourseTemplates and add some links to Canvas.
For initial launch of the whole Course Management feature / flow,
the guidance is that we setup Course Templates and use New Program Launch.
After launch, the Courses are snapshots and we're not supporting any editing of
it. In the future, we may add the ability to edit a course after launch, but all
of the things need to be thought through so this is a fail-safe to make sure
we don't shoot ourselves in the foot.