More fixing of specs and not breaking the old stuff.
We're going to do a followup PR with the core logic of
create, update, delete for a BaseCourseCustomContentVersion.

Specs are still failing. Decision is to rip out anything failing related
to New or Edit of a Course. But we need to figure out what to do with
New Course Template related failing specs.