Fix up Content Editor styling a bit.
Starts to use Bootstrap to lay things out. I didn't go the full
way and fix it all up. Just the header I did as it's own container
and the main WYSIWYG area. I left the right sidebar with the widgets
hardcoded with an absolute 300px width pushed to the right with a margin.