Remove obsolete accounts (signup/sign-in/etc) styles.
When we didn't have Bootstrap installed, we manually yanked a bunch
of Bootstrap styles from the old Join server so that the existing
HTML would work. This yanks those out.

Also, we used to have a navigation menu at the top with a burger / hot-dog
menu. That's been long-gone so rip that stuff out too.