CSS boggles the mind. Get rid of "margin" between body and first element.
Ran into this: https://stackoverflow.com/questions/7742720/css-margin-pushing-the-body-element
which was causing a scrollbar when the content was supposed to just fit.

Notes:
Just making sure that all the root divs in the body had a top margin of
0 wasn't enough. I tinkered with other things like transparent 1px borders
on body, etc. Finally, the `overflow: auto` is what seems to have done it.

We shouldn't add a top and bottom margin anyway in the root bv-lti-container.
Maybe add padding in the future if that looks better.