Get rid of FormAssembly layout.
It didn't make sense to have a new dedicated layout for FormAssembly.
The only different from the `lti_canvas` layout is that in the `new`
action we want the width to be 100%.

I also refactored the WaiversSubmissionController to pull things that were
not waivers specific into a base controller so that we can use it for
the Pre and Post-Accelerator forms (or any other FormAssembly forms). This
was a quick and dirty pass and could probably be cleaner.