Fix specs - login form issues.
Now that the footer is positioned at the bottom using absolute
coordinates, it's div can overlap with the div's before it. Position
it behind them so that the main div is clickable in Selenium tests.

Also, this form used to be on a server that would call into another server
so it was visibly slow at times. Now it authenticates against the local database
so we shouldn't need a "wait" indicator.