Back out changes to project_submission#new.
We had decided to create the new submission when the first answer came in.
I conflated issues when making the previous change. It was failing for a
different reason and also complaining about something that wasn't really a problem.
This goes back to the original implementation of project_submission#new
and gets rid of the console log about no project submission when trying to
pre-fill. We don't need to prefil if not answers have already been saved to a submission