Fix TA Sync To LMS. (#521)
B/c of some hacky stuff to deal with Platform Sections in non-Highlander
code, we can't use the same section names across different courses when
syncing a user. We don't care what section the TAs are in, so just make
sure the names are different when addign them to each course.